pyproject.toml
[tool.poetry]
name = "first_prj"
version = "0.1.0"
description = "FastAPI project"
authors = ["Ivan Ivanov <ivan@example.com>"]

packages = [
    { include = "first_prj" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.111.1"
uvicorn = {extras = ["standard"], version = "0.23.2"}
sqlalchemy = "2.0.20"
pydantic = "2.11.7"
python-dotenv = "1.0.1"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
alembic = "1.11.1"
fastapi-mail = "^1.5.0"
fastapi-limiter = "^0.1.4"
redis = "^5.3.1"
stripe = "^12.4.0"
itsdangerous = "^2.1.2"
python-multipart = "^0.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "<4"
Pillow = "^10.0.0"
psycopg2-binary = "^2.9.9"
aioredis = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

